feat: implement smith_waterman using a generic function (#6)
* feat: load benchmark data from $FRIZBEE_BENCH_PATH

...instead of from a hardcoding. This will make it easier for different
people to run the benchmarks without modifying the code and recompiling.

The original path is still the default :)

* feat: implement smith_waterman using a generic function

...instead of generating the implementation using a macro.
I've kept the proc macro around for use in benchmarks, but it is not
used in the main lib so it can be deleted safely in the future.

`Bucket` is now a trait, so that the implementation in FixedWidthBucket
can be generic over the string width. This means that Bucket no longer
stores function pointers.

All of the constants were converted to `u8` because `u16` to `u8`
conversion using into is not possible.

* refactor: add some type annotations to make rust-analyzer stop complaining

+++ src/const.rs
@@ -7,23 +7,23 @@ pub const SIMD_WIDTH: usize = 16;
 // NOTE: be vary carefuly when changing these values since they affect what can fit in
 // the u8 scoring without overflowing
 
-pub const MATCH_SCORE: u16 = 7;
-pub const MISMATCH_PENALTY: u16 = 4; // -4
-pub const GAP_OPEN_PENALTY: u16 = 3; // -3
-pub const GAP_EXTEND_PENALTY: u16 = 1; // -1
+pub const MATCH_SCORE: u8 = 7;
+pub const MISMATCH_PENALTY: u8 = 4; // -4
+pub const GAP_OPEN_PENALTY: u8 = 3; // -3
+pub const GAP_EXTEND_PENALTY: u8 = 1; // -1
 
 // bonus for matching the first character of the haystack
-pub const PREFIX_BONUS: u16 = 6;
+pub const PREFIX_BONUS: u8 = 6;
 // bonus for matching character after a delimiter in the haystack (e.g. space, comma, underscore, slash, etc)
-pub const DELIMITER_BONUS: u16 = 4;
+pub const DELIMITER_BONUS: u8 = 4;
 // bonus for matching a letter that is capitalized on the haystackA
 // FIXME: temporarily disabled until we can apply only when the char before the capital is
 // lowercase
-pub const CAPITALIZATION_BONUS: u16 = 0;
-pub const MATCHING_CASE_BONUS: u16 = 1;
+pub const CAPITALIZATION_BONUS: u8 = 0;
+pub const MATCHING_CASE_BONUS: u8 = 1;
 // bonus multiplier for the first character of the needle
-pub const FIRST_CHAR_MULTIPLIER: u16 = 1;
+pub const FIRST_CHAR_MULTIPLIER: u8 = 1;
 // bonus for haystack == needle
-pub const EXACT_MATCH_BONUS: u16 = 4;
+pub const EXACT_MATCH_BONUS: u8 = 4;
 
 // TODO: bonus for a full continuous match without gaps?
